pyproject.toml
[tool.poetry]
name = "hubmap-pancreas-data-explorer"
version = "0.1.0"
description = "Visualize pancreas data from the HuBMAP project"
authors = ["James Labyer"]
license = "BSD 3-Clause License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pandas = {extras = ["excel"], version = "^2.2.3"}
numpy = "~2.1.3"
dash = "~2.18.2"
dash-ag-grid = "~31.3.0"
plotly = "~5.24.1"
dash-bootstrap-components = "~1.6.0"
gunicorn = "~23.0.0"
pywavefront = "~1.3.3"
pillow = "~11.0.0"
ruff = "~0.8.2"
nh3 = "^0.2.20"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.config]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "~8.3.4"
pytest-cov = "~6.0.0"
bandit = "^1.8.0"

[tool.poetry.group.config.dependencies]
filetype = "^1.2.0"
python-calamine = "^0.3.1"
flask-login = "^0.6.3"
python-dotenv = "^1.0.1"
python-magic = "^0.4.27"

[tool.poetry.plugins.dotenv]
ignore="false"
location=".env"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.per-file-ignores]
# Ignore unused imports in test files, which have to import app.app for tests to work even though app.app is not directly referenced in the file
"app/tests/test_proteomics.py" = ["F401"]
"app/tests/test_ocpage.py" = ["F401"]
"app/tests/test_oc-links.py" = ["F401"]
"app/tests/test_home.py" = ["F401"]
"app/tests/test_3d.py" = ["F401"]
"app/tests/test_spatialmap.py" = ["F401"]

# Ignore out of order imports when sys.path must be updated to perform the import
"app/app.py" = ["E402"]
"config_portal/app.py" = ["E402"]
